requirements.txt
boto     == 2.38.0
lxml     == 3.4.4
pyyaml   == 3.11
requests == 2.7.0
